use long instead of BigInteger, i dont know what i was thinking;

+++ AOC2020/Day3.cs
@@ -55,7 +55,7 @@ namespace AOC2020
                 return map;
             }
 
-            public BigInteger TraverseMapAndCount(int right, int down)
+            public long TraverseMapAndCount(int right, int down)
             {
                 var i = right;
                 var treeCount = 0;
